Add files via upload
//working of program

The program begins with the main() function.
Inside main(), variables are initialized as follows: a = 10 and b = 20.
The max(a, b) function is invoked.
Inside the max() function:
It checks the values of x = 10 against y = 20.
Since y > x, it outputs value of y. So value is 20.
The return value is passed to the variable m.
Now the program is set to print m is 20 to the console.
The program will successfully complete by returning 0 from main().